Issue1304

+++ sql/components/strongComponents.sql
@@ -30,7 +30,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 CREATE OR REPLACE FUNCTION pgr_strongComponents(
     TEXT, -- edges_sql (required)
 
-    OUT seq INTEGER,
+    OUT seq BIGINT,
     OUT component BIGINT,
     OUT node BIGINT)
 RETURNS SETOF RECORD AS
